Move service&registry to data-plane/model
