+++ edc-extension4aas/src/main/java/de/fraunhofer/iosb/app/model/aas/registry/RegistryRepository.java
@@ -16,7 +16,8 @@
 package de.fraunhofer.iosb.app.model.aas.registry;
 
 import de.fraunhofer.iosb.app.model.aas.AasProviderRepository;
-import de.fraunhofer.iosb.app.model.aas.service.Service;
+import de.fraunhofer.iosb.model.aas.registry.Registry;
+import de.fraunhofer.iosb.model.aas.service.Service;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
