+++ edc-extension4aas/src/main/java/de/fraunhofer/iosb/app/model/aas/service/ServiceRepositoryUpdater.java
@@ -18,6 +18,7 @@ package de.fraunhofer.iosb.app.model.aas.service;
 import de.fraunhofer.iosb.app.pipeline.PipelineResult;
 import de.fraunhofer.iosb.app.pipeline.PipelineStep;
 import de.fraunhofer.iosb.app.util.Pair;
+import de.fraunhofer.iosb.model.aas.service.Service;
 import org.eclipse.edc.connector.controlplane.asset.spi.domain.Asset;
 
 /**
